Lecture 4 content part B
Here we have the rest of the lecture content, this includes exceptions
and preprocessing. There is a lot of general backgorund knowledge which
can be found within the files.

very breif information is recorded on each of the debugging features for
memory leaks and time as these are extensions that i dont expect to be
within the course

also note that makefiles do not work within windows